SWARM-1631: Move metrics CDI extension into fraction
