+++ fractions/microprofile/microprofile-metrics/module.conf
@@ -1,7 +1,10 @@
+javax.enterprise.api
+javax.annotation.api
+javax.api
+
 org.wildfly.swarm.logging
 org.wildfly.swarm.undertow
 org.jboss.logging
-javax.api
 
 org.eclipse.microprofile.metrics export=true
 
@@ -16,7 +19,6 @@ org.jboss.as.controller
 org.jboss.as.server
 org.jboss.as.naming
 javax.ws.rs.api
-javax.api
 
 io.undertow.core
 org.jboss.xnio
