+++ fractions/microprofile/microprofile-metrics/src/main/java/org/wildfly/swarm/microprofile/metrics/MicroprofileMetricsFraction.java
@@ -19,12 +19,11 @@ package org.wildfly.swarm.microprofile.metrics;
 import org.wildfly.swarm.spi.api.Fraction;
 import org.wildfly.swarm.spi.api.annotations.Configurable;
 import org.wildfly.swarm.spi.api.annotations.DeploymentModule;
-import org.wildfly.swarm.spi.api.annotations.DeploymentModule.MetaInfDisposition;
 
 /**
  * @author Heiko W. Rupp
  */
-@DeploymentModule(name = "org.wildfly.swarm.microprofile.metrics.cdi", metaInf = MetaInfDisposition.IMPORT)
+@DeploymentModule(name = "org.wildfly.swarm.microprofile.metrics", slot = "deployment", export = true, metaInf = DeploymentModule.MetaInfDisposition.IMPORT)
 @Configurable("swarm.microprofile.metrics")
 public class MicroprofileMetricsFraction implements Fraction<MicroprofileMetricsFraction> {
 }
