+++ fractions/microprofile/microprofile-metrics/src/main/java/org/wildfly/swarm/microprofile/metrics/deployment/TimedInterceptor.java
@@ -15,12 +15,12 @@
  *   limitations under the License.
  *
  */
-package org.wildfly.swarm.microprofile.metrics.cdi;
+package org.wildfly.swarm.microprofile.metrics.deployment;
 
 
-import org.eclipse.microprofile.metrics.Meter;
 import org.eclipse.microprofile.metrics.MetricRegistry;
-import org.eclipse.microprofile.metrics.annotation.Metered;
+import org.eclipse.microprofile.metrics.Timer;
+import org.eclipse.microprofile.metrics.annotation.Timed;
 
 import javax.annotation.Priority;
 import javax.enterprise.inject.Intercepted;
@@ -35,10 +35,10 @@ import java.lang.reflect.AnnotatedElement;
 import java.lang.reflect.Member;
 
 @SuppressWarnings("unused")
-@Metered
+@Timed
 @Interceptor
 @Priority(Interceptor.Priority.LIBRARY_BEFORE + 10)
-/* packaged-private */ class MeteredInterceptor {
+/* package-private */ class TimedInterceptor {
 
     private final Bean<?> bean;
 
@@ -47,35 +47,39 @@ import java.lang.reflect.Member;
     private final MetricResolver resolver;
 
     @Inject
-    private MeteredInterceptor(@Intercepted Bean<?> bean, MetricRegistry registry) {
+    private TimedInterceptor(@Intercepted Bean<?> bean, MetricRegistry registry) {
         this.bean = bean;
         this.registry = registry;
         this.resolver = new MetricResolver();
     }
 
     @AroundConstruct
-    private Object meteredConstructor(InvocationContext context) throws Exception {
-        return meteredCallable(context, context.getConstructor());
+    private Object timedConstructor(InvocationContext context) throws Exception {
+        return timedCallable(context, context.getConstructor());
     }
 
     @AroundInvoke
-    private Object meteredMethod(InvocationContext context) throws Exception {
-        return meteredCallable(context, context.getMethod());
+    private Object timedMethod(InvocationContext context) throws Exception {
+        return timedCallable(context, context.getMethod());
     }
 
     @AroundTimeout
-    private Object meteredTimeout(InvocationContext context) throws Exception {
-        return meteredCallable(context, context.getMethod());
+    private Object timedTimeout(InvocationContext context) throws Exception {
+        return timedCallable(context, context.getMethod());
     }
 
-    private <E extends Member & AnnotatedElement> Object meteredCallable(InvocationContext context, E element) throws Exception {
-        String name = resolver.metered(bean.getBeanClass(), element).metricName();
-        Meter meter = (Meter) registry.getMetrics().get(name);
-        if (meter == null) {
-            throw new IllegalStateException("No meter with name [" + name + "] found in registry [" + registry + "]");
+    private <E extends Member & AnnotatedElement> Object timedCallable(InvocationContext context, E element) throws Exception {
+        String name = resolver.timed(bean.getBeanClass(), element).metricName();
+        Timer timer = (Timer) registry.getMetrics().get(name);
+        if (timer == null) {
+            throw new IllegalStateException("No timer with name [" + name + "] found in registry [" + registry + "]");
         }
 
-        meter.mark();
-        return context.proceed();
+        Timer.Context time = timer.time();
+        try {
+            return context.proceed();
+        } finally {
+            time.stop();
+        }
     }
 }
