+++ fractions/microprofile/microprofile-metrics/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,0 +1,1 @@
+org.wildfly.swarm.microprofile.metrics.deployment.MetricCdiInjectionExtension
