--- fractions/microprofile/microprofile-metrics/src/main/resources/modules/org/wildfly/swarm/microprofile/metrics/cdi/main/module.xml
@@ -1,31 +1,0 @@
-<!--
-  ~ /*
-  ~   Copyright 2017 Red Hat, Inc, and individual contributors.
-  ~
-  ~   Licensed under the Apache License, Version 2.0 (the "License");
-  ~   you may not use this file except in compliance with the License.
-  ~   You may obtain a copy of the License at
-  ~
-  ~   http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~   Unless required by applicable law or agreed to in writing, software
-  ~   distributed under the License is distributed on an "AS IS" BASIS,
-  ~   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~   See the License for the specific language governing permissions and
-  ~   limitations under the License.
-  ~ */
-  -->
-
-<module xmlns="urn:jboss:module:1.3" name="org.wildfly.swarm.microprofile.metrics.cdi">
-  <resources>
-    <artifact name="org.wildfly.swarm:metrics-api:${project.version}"/>
-  </resources>
-  <dependencies>
-    <module name="javax.enterprise.api"/>
-    <module name="javax.annotation.api"/>
-    <module name="javax.api"/>
-    <module name="org.eclipse.microprofile.metrics"/>
-    <module name="org.wildfly.swarm.microprofile.metrics" slot="api"/>
-    <module name="org.jboss.logging"/>
-  </dependencies>
-</module>
